+++ src/org/usfirst/frc330/subsystems/Turret.java
@@ -433,7 +433,7 @@ public class Turret extends Subsystem implements LiveWindowSendable {
 					}
 				}
 			} else if (key.equals("Mode")) {
-				if (turret.getControlMode().getValue() != ((Integer) value).intValue()) {
+				if (turret.getControlMode().getValue() != ((Double) value).intValue()) {
 					changeControlMode(TalonControlMode.valueOf(((Integer) value).intValue()));
 				}
 			} else if (key.equals("Value")) {
